backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.8.0
google-api-python-client>=2.111.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0
mutagen>=1.47.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
